Making a mess of Prawn by starting work on these TextBox things.
Actually my aim is to make it easy to render simple tables, but for now there's TextBox, which intrudes but then does not intrude enough. The idea is sort of similar to the bounding boxes, except this is only for text and changing a TextBox's width changes its height, kind of like in HTML.

In addition to introducing the Prawn::TextBox class (which adds nothing to the namespace you'd use when doing pdf.whatever, but requires you to use an ugly "render_on_pdf" call) I moved greedy_wrap into AFM, but then couldn't get it to work right, so I rewrote it as naive_wrap. I think it does more or less the same thing.