Prawn#text_box returns excess text.
This is a fix to ticket #2, item #1: When Prawn#text_box is called,
the text which is not displayed in the text box is returned.

Note that because naive_wrap is used to fit text into the text box,
whitespace formatting (especially spaces and tabs) is not preserved in
the excess text returned; however, naive_unwrap does attempt to
restore the original line breaks.