Remove Prawn::Reference on_encode block argument, in favor of a new, more flexible callback system. Closes #56.
Added a callback system Prawn::Document::Internals#before_render(&block),
which accepts a Proc to be run after all objects are finalized, but before
anything is rendered to the output stream. This removes the need for the
on_encode Proc argument to Prawn::Reference, so that argument has been
removed.

Since we are no longer storing Procs in references, this closes issue 56.
Added a spec to verify.