Omit /Encoding from font dict for AFM symbol fonts
AFM symbolic fonts (Symbol and ZapfDingbats) have their own encodings,
and based on my best reading of the PDF 1.7 spec, should not have an
/Encoding entry in their font dictionary if using the stock encoding.
This removes that entry from the dictionary for symbolic fonts (AFM
fonts with a CharacterSet of Special are assumed symbolic).

We should be generating a full font descriptor even for the Base 14
fonts (see issue #71), but we are not there yet. This fixes the
ZapfDingbats issue reported on the list, though.