Finish out table rows before pagebreak
When a cell is taller than the margin box of a page, it would cause
subsequent columns to be rendered on the next page. This fixes things so
that the columns will be rendered next to the overflowing column.

Also ensure that we only flow to the next page to start a table if we
could gain height by doing so (i.e., if we are below the top of the
bounding box).

Thanks to TylerRick on github for the bug report and the patch that
inspired this fix.

Closes #148.