Defer cell min_width calculation until needed
min_width was set in the Prawn::Table::Cell initializer and so was
ignoring any changes that happened after the cells were built (such as
in the block passed to Prawn::Table.new). So tables with a small width
would still raise CannotFit errors, even if you tried to shrink the size
or padding in the initializer to make it work.

This fix defers min / max width calculations until they are requested,
which only happens when cells are actually being positioned.

Thanks to Liviu for reporting this issue with a simple test case.