Draw borders before content
This reverses 8e0de9a, which I probably had a good reason for at the
time but for some reason didn't document. The behavior introduced by
that commit broke table cells inside stretchy bounds, because the
absolute location of the y-position depends on when the bounds are
stretched (#153).

Introduced a spec to cover the new behavior and prevent regressions.
Closes #153.