Start long cells on same page if at top of stretchy bbox
Previously, we would use the reference bounds (margin box if the parent
bounds were stretchy) to determine whether to flow a cell that would not
fit on the current page to the next page. This incorrectly handled the
case where the parent bounds were stretchy but we were at the top. That
has been specced and fixed.