Ensure Windows-1252 is a known encoding
WriteExcel, when running in Ruby 1.8, defines an encoding-compatibility
layer; however it only supports a small handful of character sets. As a
result, when WriteExcel and Prawn are running together, the respond_to
would pass but an exception would be raised on the unknown Windows-1252
character set. This patch enables compatibility with WriteExcel and
(hopefully) any other half-assed encoding compatibility system for 1.8.