(This is just an insignificant change to the previous commit the actual bug fix is:)
Commit:2de9c03aa122ed8d16a4c542bc461ad3902048f5

* This is a fix to a bug in lib/prawn/text.rb.
Indented text followed by two returns was causing a
new page because of an empty paragraph (first para,
empty para, next para). The logic tried to set the
empty paragraph, saw that remaining text was still
the same as original paragraph, so concluded it had
not fit on the page.

The fix here is simple but not optimal as it
involves replacing the empty paragraph with a single
non-breaking space. It would be better for the logic
to handle empty paragraphs directly by simply moving
down by the line space and then going on to the next
paragraph.