Allow Pathname objects to be passed to image()
The comparison in detect_image_format was being performed against a
string in the source encoding of lib/prawn/images.rb. This is not
resilient against different input encodings of the image file.

This commit changes detect_image_format to do a byte-by-byte compare to
detect the JPEG or PNG header.