Draw all table backgrounds before any borders
Within each page of a table, all backgrounds should be drawn before any
borders. This ensures that the background of one cell doesn't overlap
the border of another.

Extracted Cell.draw_cells, which draws a collection of cells at
specified positions on the page.

Thanks to mrkn (Kenta Murata) for the bug report and initial patch.

Fixes #226.