Don't restore y-position on stretchy bboxes
When starting a bounding box at the bottom of a page (for example, using
span starting at the end of a page), drawing some text (advancing to the
next page), and then closing out the bbox / span, Prawn would restore
the original (bottom) y-position, thus stretching the bbox to the height
of the page. This commit resolves that issue.

Closes #255.