Pass through border and padding options from span dummy cells to master
This allows the border color/width and padding to propagate up from
cells that were styled as part of a colspan or rowspan (but not
explicitly part of the data themselves). Care is taken not to propagate
these values from cells or borders interior to the span (such as the
left-hand border of a rightmost cell in a colspan).

Closes #329.