Added option to ColumnBox to reset the top margin
This allows for page title text on the first page that is not repeated
on every page. The default is off, so it is completely backward
compatible. I also added a spec to verify the existing behavior
(which was not previously tested) and a spec to for the new option.