Remove transactions and temporarily disable group
It does not seem likely that we will be able to stabilize
transactions in a general way any time in the near future.
I would like to restore a somewhat limited version of
the group() feature that does not rely on transactions,
but until we do, disabling the feature will make the
problem obvious right away to users rather than much
later via some hard-to-debug symptom.