Lock to a specific version of Rubocop
The reason we lock to a specific version instead of pessimistic versions
is we don't want a new cop to appear in the test suite that will fail
previously passing code. When we upgrade Rubocop ourselves we can see
what new cops where added and enable/disable them at that time.