Use extracted prawn-manual_builder code
The functionality of the manual should stay the same,
we now just depend on an external gem rather than bundling
the manual building code directly with Prawn.

In theory, this will allow for the manual building code
to be used by other extensions. In practice, some
extra effort may be needed before that's practical.