Adding a manual entry for alternate page numbers
It was unclear how to do this, and I had submitted a previous pull
request that did not explore the existing API.

After a bit of work, this was the method that I ended up using. I
wanted to make sure that the manual (which is very helpful) was
updated.