Fix styled width_of when using TTFs.
Previously, width_of calculations on styled text were relying on the
document font's `name` attribute in order to look up the appropriate
font style. This doesn't work for TTF fonts, since the name is a full
path to a single style of font, and the Prawn must know about the font
family in order to find another style.

This update uses the family instead, which allows the width_of
calculations to work properly with TTFs, and doesn't appear to cause any
regressions based on existing specs.