Enable manual to be built reproducibly
In order to make the manual build reproducibly, we need a way to stop
writing the current build time in it.
The following patch will allow this in two ways: if BUILD_DATE is set
in the environment, the date will be taken from there; otherwise, if
the Git index is available, the date of the last commit will be used.