Skip to next JPEG segment in case of corrupt header length
JPEG segments must be `0xFF` followed by anything but `0x00`. `0xFF` are
escaped with `0xFF` in the header, so we can be sure that `0xFF0xXX` is
a segment control.

In case an image has an invalid header segment such as an incorrect
length, prawn will reject the image while other readers can account for this
error and still render the content fine.