Turn on RSpec color output on CI
I've found color output super useful reading CI output to spot what
failed. `.rspec` has `--color`, but this will only turn on color if it
can detect that the terminal supports it. Unfortunately, it determines this by checking if it's a TTY, which
GitHub Actions would not be.

This explicitly turns on color when the CI environment variable is
present.

References:

- CI environment variable: https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables
- Color support: https://github.blog/2020-09-23-a-better-logs-experience-with-github-actions/