Allow URLs as images
Using Prawn in an environment where assets are served
dynamically can be painful. This aims to solve it.

For instance, it is convenient to use Rails' ActionController
helper asset_path to retrieve a specific image. Trying to come
up with rules to assess the asset relative path looks brittle.
Precisely, frameworks like the Asset Pipeline solves this.
However, Prawn lacks support for these scenarios.