Full font embedding
This add an option to disable font subsetting. Original fonts can be
embedded in full original form.

This feature can make documents substantially bigger. In addition to
embedded fonts being bigger PDF requires additional information in order
to properly render text. Specifically, it requires glyph widths. Some
fonts contain thousands of glyps. A thousand of glyph widths on average
would result in about 4 Kb additional size of the document.
Additionally, PDF requires another mapping to make the text intelligible
when copying. This additional size is much harder to estimate as it
greatly depend on the font coverage but usually on the order of ~1-10
Kb per font.

Intended use case is a workaround for when TTFunk breaks fonts in
subsetting. But also this might be useful for documents that are going
to be edited. For example, documents that are templates and more text
would be added later, or AcroForm feature that allows end users to fill
forms.